requirements.txt
python-masscan
requests
